fastapi-project/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pandas==2.0.1
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==0.8.11
pydantic==1.10.7
python-dotenv==1.0.0